Damage values, minor config cleanup, prepare plate texture.
