+++ addons/arcadian/base/config_plateinfo.hpp
@@ -1,0 +1,7 @@
+class PlateInfos {
+    name = "spz";
+    color[] = {0, 0, 0, 0.75};
+    plateFont = "RobotoCondensedLight";
+    plateFormat = "TACS - ##"; // Will show as TACS - 01 in game.
+    plateLetters = "ABCDEFHIKLMOPRSTVXYZ";
+};
