+++ addons/arcadian/data/arcadian_armour_plate.bisurf
@@ -2,7 +2,7 @@ Density = 8600;
 Thickness = 16;
 rough = 0.1;
 dust = 0;
-bulletPenetrabilityWithThickness = 16;
+bulletPenetrabilityWithThickness = 16.5;
 soundEnviron = "Empty";
 isWater = "false";
 friction = 0.69;
