+++ addons/arcadian/data/arcadian_body_armed.rvmat
@@ -7,7 +7,7 @@ specularPower = 500;
 PixelShaderID = "Super";
 VertexShaderID = "Super";
 class Stage1 {
-    texture = "x\tacs\addons\arcadian\data\arcadian_body_armed.paa";
+    texture = "x\tacs\addons\arcadian\data\arcadian_body_armed_nohq.paa";
     uvSource = "tex";
     class uvTransform {
         aside[] = {1,0,0};
