+++ addons/arcadian/model.cfg
@@ -384,7 +384,7 @@ class CfgModels {
                 source = "HitGlass1";
                 minValue = 0;
                 maxValue = 1;
-                hidevalue = 0.99;
+                hidevalue = 0.9999;
             };
             class Glass2_destruct: Glass1_destruct {
                 selection = "glass2";
@@ -397,6 +397,22 @@ class CfgModels {
             class Glass4_destruct: Glass1_destruct {
                 selection = "glass4";
                 source = "HitGlass4";
+            };
+            class Glass5_destruct: Glass1_destruct {
+                selection = "glass5";
+                source = "HitGlass5";
+            };
+            class Glass6_destruct: Glass1_destruct {
+                selection = "glass6";
+                source = "HitGlass6";
+            };
+            class Glass7_destruct: Glass1_destruct {
+                selection = "glass7";
+                source = "HitGlass7";
+            };
+            class Glass8_destruct: Glass1_destruct {
+                selection = "glass8";
+                source = "HitGlass8";
             };
             class IndicatorSpeed {
                 animPeriod = 0;
@@ -443,22 +459,6 @@ class CfgModels {
             "podsvit pristroju"
         };
         class Animations: Animations {
-            class Glass5_destruct: Glass1_destruct {
-                selection = "glass5";
-                source = "HitGlass5";
-            };
-            class Glass6_destruct: Glass1_destruct {
-                selection = "glass6";
-                source = "HitGlass6";
-            };
-            class Glass7_destruct: Glass1_destruct {
-                selection = "glass7";
-                source = "HitGlass7";
-            };
-            class Glass8_destruct: Glass1_destruct {
-                selection = "glass8";
-                source = "HitGlass8";
-            };
             class wheel_1_1_Damage: wheel_1_1_Damage {
                 offset1 = 0.18;
             };
@@ -577,37 +577,37 @@ class CfgModels {
                 angle1 = rad 80;
             };
             class ram_hide_source {
-                type="hide";
-                source="ram_hide_source";
-                selection="ram_hide";
-                maxValue=1;
-                minValue=0;
-                hideValue=1;
+                type = "hide";
+                source = "ram_hide_source";
+                selection = "ram_hide";
+                maxValue = 1;
+                minValue = 0;
+                hideValue = 1;
             };
             class antenna_hide_source {
-                type="hide";
-                source="antenna_hide_source";
-                selection="antenna_hide";
-                maxValue=1;
-                minValue=0;
-                hideValue=1;
+                type = "hide";
+                source = "antenna_hide_source";
+                selection = "antenna_hide";
+                maxValue = 1;
+                minValue = 0;
+                hideValue = 1;
             };
             class roofbar_hide_source {
-                type="hide";
-                source="roofbar_hide_source";
-                selection="roofbar_hide";
-                maxValue=1;
-                minValue=0;
-                hideValue=1;
+                type = "hide";
+                source = "roofbar_hide_source";
+                selection = "roofbar_hide";
+                maxValue = 1;
+                minValue = 0;
+                hideValue = 1;
             };
             class roofbarexternal_hide_source: roofbar_hide_source {
-                type="hide";
-                source="roofbar_hide_source";
-                selection="roofbarexternal_hide";
-                maxValue=1;
-                minValue=0;
-                unhideValue=1;
-                hideValue=0;
+                type = "hide";
+                source = "roofbar_hide_source";
+                selection = "roofbarexternal_hide";
+                maxValue = 1;
+                minValue = 0;
+                unhideValue = 1;
+                hideValue = 0;
             };
         };
     };
